requirements.txt
alembic==1.9.0
# bcrypt==3.2.0
bcrypt==4.0.1
cffi==1.15.0
click==8.1.3
dnspython==2.1.0
email-validator==1.1.3

# Flask==2.0.1
# Flask-Bcrypt==0.7.1
# Flask-Login==0.5.0
# Flask-Migrate==3.1.0
# Flask-SQLAlchemy==2.5.1
# Flask-WTF==0.15.1
# SQLAlchemy==1.4.23
# Werkzeug==2.0.1

Flask==2.3.3
Flask-Bcrypt==1.0.1 
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-Session==0.5.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.1.1
Flask-SocketIO==5.3.2
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9

Werkzeug==2.3.7 
gunicorn==21.2.0
requests==2.31.0

pypdf==3.15.0
pdf2image==1.16.3
openpyxl==3.1.2
dropbox==11.36.2

langchain==0.0.354
langchain-community==0.0.8
langchain-core==0.1.6
langchainhub==0.1.14
langsmith==0.0.77
tiktoken==0.5.2

requests==2.31.0
requests-oauthlib==1.3.1
types-requests==2.31.0.2
lxml==5.0.0


chromadb==0.4.22

greenlet==1.1.1
idna==3.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.1.5
MarkupSafe==2.1.1
pycparser==2.20
python-dateutil==2.8.2
python-dotenv==1.0.0
python-editor==1.0.4
six==1.16.0
WTForms==2.3.3